made the map iterator no longer a parallel stream, renamed mixin to mixins, moved MixinHooks to the core package
