+++ src/main/java/io/github/darealturtywurty/turtychemistry/common/block/entity/AnvilBlockEntity.java
@@ -29,7 +29,7 @@ public final class AnvilBlockEntity extends ModularBlockEntity {
             setStackInSlot(new ItemStack(ItemInit.Sheets.COPPER_SHEET.get()));
             return;
         }
-        ItemInit.INGOT_SHEET_MAP.entrySet().parallelStream().forEach(entry -> {
+        ItemInit.INGOT_SHEET_MAP.entrySet().forEach(entry -> {
             if (getItem().is(entry.getKey().get())) {
                 setStackInSlot(new ItemStack(entry.getValue().get()));
             }
