+++ src/main/java/io/github/darealturtywurty/turtychemistry/core/MixinHooks.java
@@ -1,0 +1,6 @@
+package io.github.darealturtywurty.turtychemistry.core;
+
+public final class MixinHooks {
+
+
+}
