+++ src/main/java/io/github/darealturtywurty/turtychemistry/core/init/ItemInit.java
@@ -10,6 +10,8 @@ import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 
+import java.util.function.Supplier;
+
 public final class ItemInit extends AbstractInit {
     public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS,
             TurtyChemistry.MODID);
@@ -58,9 +60,10 @@ public final class ItemInit extends AbstractInit {
     public static final RegistryObject<ChemistryItem> RUBBER = ITEMS.register("rubber",
             () -> new ChemistryItem(makeItemProperties()));
     public static final RegistryObject<ChemistryItem> BASIC_HAMMER = ITEMS.register("basic_hammer",
-            () -> new BasicHammer(makeItemProperties()));
+            () -> new BasicHammer(new ChemistryItem.Builder(
+                    new Item.Properties().tab(TurtyChemistry.TAB).durability(70).stacksTo(1))));
     public static final ImmutableMap<Supplier<Item>, Supplier<Item>> INGOT_SHEET_MAP = ImmutableMap.<Supplier<Item>, Supplier<Item>>builder()
-            .put(Ingots.ALUMINUM_INGOT, Sheets.ALUMINUM_SHEET).build();
+            .put(Ingots.ALUMINUM_INGOT::get, Sheets.ALUMINUM_SHEET::get).build();
 
     static {
         ItemInit.Elements.register();
