--- src/main/java/io/github/darealturtywurty/turtychemistry/mixin/MixinHooks.java
@@ -1,6 +1,0 @@
-package io.github.darealturtywurty.turtychemistry.mixin;
-
-public final class MixinHooks {
-
-
-}
