+++ src/main/resources/turtychemistry.mixins.json
@@ -1,6 +1,6 @@
 {
 	"required": true,
-	"package": "io.github.darealturtywurty.turtychemistry.mixin",
+	"package": "io.github.darealturtywurty.turtychemistry.mixins",
 	"compatibilityLevel": "JAVA_17",
 	"refmap": "turtychemistry.refmap.json",
 	"mixins": [
